Prettier fix

+++ contracts/TalentLayerEscrow.sol
@@ -130,7 +130,7 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @notice Emitted after the origin platform fee was updated
      * @param _originPlatformFee The new origin platform fee
      */
-    event OriginPlatformFeeUpdated(uint256 _originPlatformFee);
+    event OriginPlatformFeeUpdated(uint16 _originPlatformFee);
 
     /**
      * @notice Emitted after a platform withdraws its balance
@@ -298,7 +298,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _talentLayerIDAddress Contract address to TalentLayerID.sol
      * @param _talentLayerPlatformIDAddress Contract address to TalentLayerPlatformID.sol
      */
-    constructor(address _serviceRegistryAddress, address _talentLayerIDAddress, address _talentLayerPlatformIDAddress) {
+    constructor(
+        address _serviceRegistryAddress,
+        address _talentLayerIDAddress,
+        address _talentLayerPlatformIDAddress
+    ) {
         serviceRegistryContract = IServiceRegistry(_serviceRegistryAddress);
         talentLayerIdContract = ITalentLayerID(_talentLayerIDAddress);
         talentLayerPlatformIdContract = ITalentLayerPlatformID(_talentLayerPlatformIDAddress);
@@ -836,7 +840,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _metaEvidence The meta evidence of the transaction
      * @param _sellerId The ID of the seller
      */
-    function _afterCreateTransaction(uint256 _transactionId, string memory _metaEvidence, uint256 _sellerId) internal {
+    function _afterCreateTransaction(
+        uint256 _transactionId,
+        string memory _metaEvidence,
+        uint256 _sellerId
+    ) internal {
         Transaction storage transaction = transactions[_transactionId];
 
         uint256 sender = talentLayerIdContract.walletOfOwner(transaction.sender);
@@ -866,7 +874,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _token The token to transfer
      * @param _amount The amount of tokens to transfer
      */
-    function _deposit(address _sender, address _token, uint256 _amount) private {
+    function _deposit(
+        address _sender,
+        address _token,
+        uint256 _amount
+    ) private {
         require(IERC20(_token).transferFrom(_sender, address(this), _amount), "Transfer must not fail");
     }
 
@@ -949,10 +961,7 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _proposalId The id of the proposal
      * @return proposal proposal struct, service The service struct, sender The sender address, receiver The receiver address
      */
-    function _getTalentLayerData(
-        uint256 _serviceId,
-        uint256 _proposalId
-    )
+    function _getTalentLayerData(uint256 _serviceId, uint256 _proposalId)
         private
         returns (
             IServiceRegistry.Proposal memory proposal,
@@ -974,10 +983,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _proposalId The id of the proposal
      * @return The Proposal struct
      */
-    function _getProposal(
-        uint256 _serviceId,
-        uint256 _proposalId
-    ) private view returns (IServiceRegistry.Proposal memory) {
+    function _getProposal(uint256 _serviceId, uint256 _proposalId)
+        private
+        view
+        returns (IServiceRegistry.Proposal memory)
+    {
         return serviceRegistryContract.getProposal(_serviceId, _proposalId);
     }
 
@@ -996,7 +1006,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _tokenAddress The token address
      * @param _amount The amount to transfer
      */
-    function _transferBalance(address payable _recipient, address _tokenAddress, uint256 _amount) private {
+    function _transferBalance(
+        address payable _recipient,
+        address _tokenAddress,
+        uint256 _amount
+    ) private {
         if (address(0) == _tokenAddress) {
             _recipient.transfer(_amount);
         } else {
@@ -1004,7 +1018,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
         }
     }
 
-    function _safeTransferBalance(address payable _recipient, address _tokenAddress, uint256 _amount) private {
+    function _safeTransferBalance(
+        address payable _recipient,
+        address _tokenAddress,
+        uint256 _amount
+    ) private {
         if (address(0) == _tokenAddress) {
             _recipient.call{value: _amount}("");
         } else {
@@ -1018,10 +1036,11 @@ contract TalentLayerEscrow is Ownable, IArbitrable {
      * @param _platformFee The platform fee
      * @return totalEscrowAmount The total amount to be paid by the buyer (including all fees + escrow) The amount to transfer
      */
-    function _calculateTotalEscrowAmount(
-        uint256 _amount,
-        uint256 _platformFee
-    ) private view returns (uint256 totalEscrowAmount) {
+    function _calculateTotalEscrowAmount(uint256 _amount, uint256 _platformFee)
+        private
+        view
+        returns (uint256 totalEscrowAmount)
+    {
         return
             _amount +
             (((_amount * protocolFee) + (_amount * originPlatformFee) + (_amount * _platformFee)) / FEE_DIVIDER);
